Repackaged Context module: * Separated DateTime functions into separate file
